+++ package.json
@@ -75,11 +75,11 @@
     "mocha": "cd packages/terra-functional-testing && npm run mocha",
     "pack": "webpack -p",
     "postinstall": "link-parent-bin && npm run compile && npm run bootstrap:hoist",
+    "prepare-for-release": "npm run terra -- prepare-for-release",
     "start": "webpack-dev-server --config webpack.dev-site.config.js",
     "start-heroku": "node ./scripts/hack-express-server.js --port $PORT --site './build' ",
     "terra": "node ./packages/terra-cli/bin/terra",
     "test": "npm run lint && npm run jest && npm run mocha && npm run wdio",
-    "yupversion": "npm run terra -- version",
     "wdio-site": "SITE=build npm run terra -- wdio",
     "wdio-webpack": "npm run terra -- wdio",
     "wdio": "rm -rf build && npm run pack && npm run wdio-site && npm run wdio-webpack"
