--- packages/terra-cli/tests/jest/fixtures/normal-project/lib/terra-cli/command-11/index.js
@@ -1,7 +1,0 @@
-const cli = {
-  command: 'command-11',
-  describe: 'Run command-11',
-  handler: () => {},
-};
-
-module.exports = cli;
