--- packages/terra-cli/tests/jest/fixtures/normal-project/lib/terra-cli/command-12/index.js
@@ -1,7 +1,0 @@
-const cli = {
-  command: 'command-12',
-  describe: 'Run command-12',
-  handler: () => {},
-};
-
-module.exports = cli;
