+++ packages/terra-cli/tests/jest/fixtures/normal-project/src/terra-cli/command-12/index.js
@@ -1,0 +1,7 @@
+const cli = {
+  command: 'command-12',
+  describe: 'Run command-12',
+  handler: () => {},
+};
+
+module.exports = cli;
