--- packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/a/lib/terra-cli/command-1/index.js
@@ -1,7 +1,0 @@
-const cli = {
-  command: 'command-1',
-  describe: 'Run command-1',
-  handler: () => {},
-};
-
-module.exports = cli;
