+++ packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/a/src/terra-cli/command-2/index.js
@@ -1,0 +1,7 @@
+const cli = {
+  command: 'command-2',
+  describe: 'Run command-2',
+  handler: () => {},
+};
+
+module.exports = cli;
