--- packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/b/lib/terra-cli/command-3/index.js
@@ -1,7 +1,0 @@
-const cli = {
-  command: 'command-3',
-  describe: 'Run command-3',
-  handler: () => {},
-};
-
-module.exports = cli;
