+++ packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/b/src/terra-cli/command-3/index.js
@@ -1,0 +1,7 @@
+const cli = {
+  command: 'command-3',
+  describe: 'Run command-3',
+  handler: () => {},
+};
+
+module.exports = cli;
