+++ packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/b/src/terra-cli/command-4/index.js
@@ -1,0 +1,7 @@
+const cli = {
+  command: 'command-4',
+  describe: 'Run command-4',
+  handler: () => {},
+};
+
+module.exports = cli;
