--- packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/c/lib/terra-cli/command-6/index.js
@@ -1,7 +1,0 @@
-const cli = {
-  command: 'command-6',
-  describe: 'Run command-6',
-  handler: () => {},
-};
-
-module.exports = cli;
