+++ packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/c/src/terra-cli/command-5/index.js
@@ -1,0 +1,7 @@
+const cli = {
+  command: 'command-5',
+  describe: 'Run command-5',
+  handler: () => {},
+};
+
+module.exports = cli;
