+++ packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/c/src/terra-cli/command-6/index.js
@@ -1,0 +1,7 @@
+const cli = {
+  command: 'command-6',
+  describe: 'Run command-6',
+  handler: () => {},
+};
+
+module.exports = cli;
