--- packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/d/lib/terra-cli/command-7/index.js
@@ -1,7 +1,0 @@
-const cli = {
-  command: 'command-7',
-  describe: 'Run command-7',
-  handler: () => {},
-};
-
-module.exports = cli;
