+++ packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/d/src/terra-cli/command-7/index.js
@@ -1,0 +1,7 @@
+const cli = {
+  command: 'command-7',
+  describe: 'Run command-7',
+  handler: () => {},
+};
+
+module.exports = cli;
