+++ packages/terra-cli/tests/jest/fixtures/terra-toolkit/packages/d/src/terra-cli/command-8/index.js
@@ -1,0 +1,7 @@
+const cli = {
+  command: 'command-8',
+  describe: 'Run command-8',
+  handler: () => {},
+};
+
+module.exports = cli;
