+++ packages/terra-cli/tests/jest/terra-cli.test.js
@@ -11,7 +11,7 @@ describe('terra-cli', () => {
   });
 
   it('returns the available commands for terra-toolkit by looking in the packages directory', async () => {
-    const { stdout } = await exec('../../../../bin/terra --help', { cwd: path.join(__dirname, 'fixtures', 'normal-project') });
+    const { stdout } = await exec('../../../../bin/terra --help', { cwd: path.join(__dirname, 'fixtures', 'terra-toolkit') });
     expect(stdout).toMatchSnapshot();
   });
 });
