--- packages/terra-open-source-scripts/src/terra-cli/version/index.js
@@ -1,11 +1,0 @@
-const versionMonoRepoHandler = require('./versionMonoRepoHandler');
-const versionRepoHandler = require('./versionRepoHandler');
-const isMonoRepo = require('../../utils/isMonoRepo');
-
-const version = {
-  command: 'version',
-  describe: 'Update the version and changelog of the project(s)',
-  handler: async () => (await isMonoRepo() ? versionMonoRepoHandler() : versionRepoHandler()),
-};
-
-module.exports = version;
